Some more changes had to be made, in order to recover from a broken stack (due to inductive discharge):
The IP-Connection has to be dropped and a new one has to be
acquired.
This is, what the new 'robust' code does.
Furthermore, all device-representations need to be updated with the new 
connection to the true Brick/let.
This is, what the new 'robust' code does.

Now, even after a discharge near the unprotected tinkerforge stack, the
software can re-establish a connection to the stack and go on working.

Finally!

Finally? 
Hard real-life tests will follow...